# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Task": {
				"properties": {
					"id": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"assignedAssistant": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"in_progress",
							"completed",
							"failed"
						]
					},
					"inputData": {
						"type": "string",
						"nullable": true
					},
					"outputData": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"description",
					"assignedAssistant",
					"status",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Task.Exclude_keyofTask.id-or-createdAt-or-updatedAt__": {
				"properties": {
					"description": {
						"type": "string"
					},
					"assignedAssistant": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"in_progress",
							"completed",
							"failed"
						]
					},
					"inputData": {
						"type": "string"
					},
					"outputData": {
						"type": "string"
					}
				},
				"required": [
					"description",
					"assignedAssistant",
					"status"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Task.id-or-createdAt-or-updatedAt_": {
				"$ref": "#/components/schemas/Pick_Task.Exclude_keyofTask.id-or-createdAt-or-updatedAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Partial_Omit_Task.id-or-createdAt-or-updatedAt__": {
				"properties": {
					"description": {
						"type": "string"
					},
					"assignedAssistant": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"in_progress",
							"completed",
							"failed"
						]
					},
					"inputData": {
						"type": "string"
					},
					"outputData": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"Partial_Omit_Tag.id__": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"Tag": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Assistant": {
				"properties": {
					"id": {
						"type": "string"
					},
					"gptAssistantId": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"chat",
							"assistant"
						]
					},
					"model": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"type",
					"model",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Partial_Assistant_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"gptAssistantId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"assistant",
							"chat"
						]
					},
					"model": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "gpt-assistants-api",
		"version": "0.2.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/task/{taskId}": {
			"get": {
				"operationId": "GetTaskById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Task"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTask",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Omit_Task.id-or-createdAt-or-updatedAt__"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTask",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/task": {
			"get": {
				"operationId": "GetAllTasks",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Task"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddTask",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_Task.id-or-createdAt-or-updatedAt_"
							}
						}
					}
				}
			}
		},
		"/task/status/{status}": {
			"get": {
				"operationId": "GetTasksByStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Task"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "status",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"pending",
								"in_progress",
								"completed",
								"failed"
							]
						}
					}
				]
			}
		},
		"/task/assistant/{assistantId}": {
			"get": {
				"operationId": "GetTasksByAssistant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Task"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tag": {
			"post": {
				"operationId": "AddTag",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"tagId": {
											"type": "string"
										}
									},
									"required": [
										"tagId"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									}
								},
								"required": [
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllTags",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Tag"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": []
			}
		},
		"/tag/{tagId}": {
			"delete": {
				"operationId": "RemoveTag",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tagId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTag",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tagId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Omit_Tag.id__"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetTagById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Tag"
								}
							}
						}
					}
				},
				"tags": [
					"Tag"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tagId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/memory-extra/{memoryId}/tags": {
			"put": {
				"operationId": "UpdateMemoryTags",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"MemoryExtra"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "memoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newTags": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"required": [
									"newTags"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/assistant/{id}": {
			"get": {
				"operationId": "GetAssistantById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Assistant"
								}
							}
						}
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAssistant",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Assistant_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAssistant",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}